refactor: UI 모델에 해당하는 클래스들 패키지 분리
