+++ app/src/main/java/com/treasurehunt/ui/model/MapSymbol.kt
@@ -1,4 +1,4 @@
-package com.treasurehunt.ui.home
+package com.treasurehunt.ui.model
 
 import kotlinx.parcelize.Parcelize
 import android.os.Parcelable
