+++ app/src/main/java/com/treasurehunt/ui/model/MapUiState.kt
@@ -1,4 +1,4 @@
-package com.treasurehunt.data.remote.model
+package com.treasurehunt.ui.model
 
 import com.naver.maps.map.overlay.Marker
 import com.treasurehunt.data.local.model.PlaceEntity
